Use memsize type for indexer for std::vector
* Eliminate possible problems with large (> INT_MAX) indexes.
* Remove potentially error-prone int -> memsize type (unsigned *)
  implicit cast.

+++ DEPENDENCIES/simple_ml_helib/src/h5Dumper.cpp
@@ -56,7 +56,7 @@ void H5Dumper::setLayerNames(const vector<string>& layerNames)
   StrType strDataType(PredType::C_S1, H5T_VARIABLE);
   vector<const char*> layerNamesCStr(layerNames.size());
 
-  for(int i = 0; i < layerNames.size(); i++)
+  for(size_t i = 0; i < layerNames.size(); i++)
   {
     layerNamesCStr[i] = layerNames[i].c_str();
   }
@@ -83,7 +83,7 @@ void H5Dumper::setLayerWeightNames(const string& layerName,
   StrType strDataType(PredType::C_S1, H5T_VARIABLE);
   vector<const char*> weightNamesCStr(weightNames.size());
 
-  for(int i = 0; i < weightNames.size(); i++)
+  for(size_t i = 0; i < weightNames.size(); i++)
   {
     weightNamesCStr[i] = weightNames[i].c_str();
   }
